Convert all floats to doubles for more precision
Although the transition from floats to doubles negatively effects cache
utilization, it's a necessary compromise for more floating point
precision. Shadow acne otherwise plagues more complex shapes.

+++ samples/red_sphere.zig
@@ -7,7 +7,7 @@ pub fn main() !void {
     const canvas_pixels = 400;
     const wall_z = 10;
     const wall_size = 7.0;
-    const pixel_size = wall_size / @intToFloat(f32, canvas_pixels);
+    const pixel_size = wall_size / @intToFloat(f64, canvas_pixels);
     const half = wall_size / 2.0;
 
     const sphere = rt.sph.sphere();
@@ -22,11 +22,11 @@ pub fn main() !void {
 
     var y: u32 = 0;
     while (y < canvas_pixels) : (y += 1) {
-        const world_y = half - pixel_size * @intToFloat(f32, y);
+        const world_y = half - pixel_size * @intToFloat(f64, y);
 
         var x: u32 = 0;
         while (x < canvas_pixels) : (x += 1) {
-            const world_x = -half + pixel_size * @intToFloat(f32, x);
+            const world_x = -half + pixel_size * @intToFloat(f64, x);
             const position = rt.tup.point(world_x, world_y, wall_z);
             const r = rt.ray.ray(ray_origin, rt.tup.normalize(position - ray_origin));
